requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
httpx
beautifulsoup4
lxml
python-dotenv
pydantic
easyocr
Pillow
python-multipart
pytest
pytest-cov
ruff
